reverse change of combining blocks purger and tenant deletion
