+++ pkg/chunk/purger/blocks_purger.go
@@ -1,7 +1,6 @@
 package purger
 
 import (
-	"context"
 	"crypto/md5"
 	"encoding/binary"
 	"encoding/hex"
@@ -9,13 +8,10 @@ import (
 	fmt "fmt"
 	"net/http"
 	"sort"
-	strings "strings"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
-	"github.com/oklog/ulid"
-	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/promql/parser"
@@ -36,7 +32,7 @@ type BlocksPurgerAPI struct {
 }
 
 func NewBlocksPurgerAPI(storageCfg cortex_tsdb.BlocksStorageConfig, cfgProvider bucket.TenantConfigProvider, logger log.Logger, reg prometheus.Registerer, cancellationPeriod time.Duration) (*BlocksPurgerAPI, error) {
-	bucketClient, err := createBucketClient(storageCfg, logger, reg)
+	bucketClient, err := createBucketClient(storageCfg, logger, "blocks-purger", reg)
 	if err != nil {
 		return nil, err
 	}
@@ -51,15 +47,6 @@ func newBlocksPurgerAPI(bkt objstore.Bucket, cfgProvider bucket.TenantConfigProv
 		logger:                    logger,
 		deleteRequestCancelPeriod: cancellationPeriod,
 	}
-}
-
-func createBucketClient(cfg cortex_tsdb.BlocksStorageConfig, logger log.Logger, reg prometheus.Registerer) (objstore.Bucket, error) {
-	bucketClient, err := bucket.NewClient(context.Background(), cfg.Bucket, "purger", logger, reg)
-	if err != nil {
-		return nil, errors.Wrap(err, "create bucket client")
-	}
-
-	return bucketClient, nil
 }
 
 func (api *BlocksPurgerAPI) AddDeleteRequestHandler(w http.ResponseWriter, r *http.Request) {
@@ -244,80 +231,4 @@ func getTombstoneHash(startTime int64, endTime int64, selectors []string) string
 
 	md5Bytes := md5.Sum(nil)
 	return hex.EncodeToString(md5Bytes[:])
-}
-
-func (api *BlocksPurgerAPI) DeleteTenant(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	userID, err := tenant.TenantID(ctx)
-	if err != nil {
-		// When Cortex is running, it uses Auth Middleware for checking X-Scope-OrgID and injecting tenant into context.
-		// Auth Middleware sends http.StatusUnauthorized if X-Scope-OrgID is missing, so we do too here, for consistency.
-		http.Error(w, err.Error(), http.StatusUnauthorized)
-		return
-	}
-
-	err = cortex_tsdb.WriteTenantDeletionMark(r.Context(), api.bucketClient, userID, api.cfgProvider, cortex_tsdb.NewTenantDeletionMark(time.Now()))
-	if err != nil {
-		level.Error(api.logger).Log("msg", "failed to write tenant deletion mark", "user", userID, "err", err)
-
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	level.Info(api.logger).Log("msg", "tenant deletion mark in blocks storage created", "user", userID)
-
-	w.WriteHeader(http.StatusOK)
-}
-
-type DeleteTenantStatusResponse struct {
-	TenantID      string `json:"tenant_id"`
-	BlocksDeleted bool   `json:"blocks_deleted"`
-}
-
-func (api *BlocksPurgerAPI) DeleteTenantStatus(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	userID, err := tenant.TenantID(ctx)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	result := DeleteTenantStatusResponse{}
-	result.TenantID = userID
-	result.BlocksDeleted, err = api.isBlocksForUserDeleted(ctx, userID)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	util.WriteJSONResponse(w, result)
-}
-
-func (api *BlocksPurgerAPI) isBlocksForUserDeleted(ctx context.Context, userID string) (bool, error) {
-	var errBlockFound = errors.New("block found")
-
-	userBucket := bucket.NewUserBucketClient(userID, api.bucketClient, api.cfgProvider)
-	err := userBucket.Iter(ctx, "", func(s string) error {
-		s = strings.TrimSuffix(s, "/")
-
-		_, err := ulid.Parse(s)
-		if err != nil {
-			// not block, keep looking
-			return nil
-		}
-
-		// Used as shortcut to stop iteration.
-		return errBlockFound
-	})
-
-	if errors.Is(err, errBlockFound) {
-		return false, nil
-	}
-
-	if err != nil {
-		return false, err
-	}
-
-	// No blocks found, all good.
-	return true, nil
 }
